requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
numpy==1.26.4
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-multipart==0.0.20
